refactor: 리뷰 반영 및 네이밍, 레이어 구조 개선
