+++ src/renderer/src/entities/shop/model/usePurchaseProduct.ts
@@ -1,14 +1,13 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { purchaseProduct } from "@/entities/shop/api/purchase";
+import { purchaseProduct, type PurchaseRequest } from "@/entities/shop/api/purchaseApi";
 
 export const usePurchaseProduct = () => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: (productId: number) => purchaseProduct(productId),
+    mutationFn: (request: PurchaseRequest) => purchaseProduct(request),
 
     onSuccess: () => {
-      // 잔액 / 프로필 / 토큰 갱신
       queryClient.invalidateQueries({ queryKey: ["user"] });
     },
   });
